refactor(stop-web-server): extracts stop web server method to own file
